render: T-071 delete legacy sun-shadow paths
The screen-space sun depth map (T-070, PR #406) is now the only sun
shadow path. Per architect direction on PR #423, deletes:

- Analytic SDF caster path: SunShadowShapeCasterBuffer,
  collectShapeCastersForCanvas + sun-cone culling, analyticShapeShadowHit,
  per-shape SDF helpers in c_compute_sun_shadow.{glsl,metal}, and the
  shapeCasterCount field of FrameDataSun.
- Legacy DDA branch: the useScreenSpaceShadow==0 block in the sun-shadow
  shaders, OccupancyEntityBounds SSBO + per-entity bounds tracker, and
  the occupancyBoundsCount field of FrameDataSun.
- Feature-flag plumbing: useScreenSpaceShadow_ from FrameDataSun,
  IRRender::setScreenSpaceShadowsEnabled / m_screenSpaceShadowsEnabled,
  and the --screen-space-shadow opt-in from shape_debug.
- SDF-rasterization step in BUILD_OCCUPANCY_GRID (PopulateSDFs block).
  This step rasterized SDF shape entities into the occupancy bitfield
  only to feed the legacy DDA. AO and light-volume only ever needed
  voxel-pool occupancy.

BAKE_SUN_SHADOW_MAP now owns the full FrameDataSun UBO upload (sun
direction, basis, AABB, flags) so downstream consumers see coherent
state every frame. resolveSun() — which scans for an ECS DIRECTIONAL
C_LightSource override — moved from system_compute_sun_shadow.hpp to
system_bake_sun_shadow_map.hpp where it's now the sole caller.

FrameDataSun shrinks from 96 to 80 bytes; std140 layout test updated.

CLAUDE.md updates: drop legacy-DDA / analytic-caster references; mark
BUILD_OCCUPANCY_GRID + OccupancyGrid SSBO as a phased-out producer
slated for removal in T-09Y after T-09X (AO migration) and T-072
(light-volume GPU port) land.

Closes #358
