+++ engine/prefabs/irreden/render/systems/system_bake_sun_shadow_map.hpp
@@ -2,20 +2,25 @@
 #define SYSTEM_BAKE_SUN_SHADOW_MAP_H
 
 // Bakes a sun-aligned 2D depth buffer from the iso canvas distance
-// texture so COMPUTE_SUN_SHADOW can replace its 64-step march with a
-// single texel read. Flag-guarded by IRRender::setScreenSpaceShadowsEnabled.
+// texture so COMPUTE_SUN_SHADOW resolves shadows with a single texel
+// read per pixel.
 //
 // Pipeline: must run after the geometry stages (trixelDistances must be
 // populated) and before COMPUTE_SUN_SHADOW (which reads the baked buffer).
+// Owns the full FrameDataSun UBO upload — sun direction, basis, AABB,
+// flags — so downstream consumers (COMPUTE_SUN_SHADOW, COMPUTE_VOXEL_AO,
+// LIGHTING_TO_TRIXEL) read coherent state every frame.
 //
-// Slot 28 is shared with the legacy occupancy grid; both systems rebind
-// it before their own dispatch.
+// Slot 28 is shared with the occupancy grid (read by AO + light-volume).
+// Both producers rebind the slot before their own dispatch, so the alias
+// is safe.
 
 #include <irreden/ir_render.hpp>
 #include <irreden/ir_system.hpp>
 #include <irreden/ir_math.hpp>
 #include <irreden/ir_profile.hpp>
 
+#include <irreden/render/components/component_light_source.hpp>
 #include <irreden/render/components/component_triangle_canvas_textures.hpp>
 #include <irreden/render/components/component_canvas_sun_shadow.hpp>
 #include <irreden/render/components/component_trixel_canvas_render_behavior.hpp>
@@ -38,11 +43,57 @@ namespace IRSystem {
 constexpr int kSunShadowMapDim = 1024;
 // Sweeps the visible AABB along -sunDir before projecting to sun-space,
 // so off-screen surfaces within this many voxels still cast into-frame
-// shadows. Mirrors the legacy march length so shadow distance survives
-// the flag flip.
+// shadows. Bounds the shadow throw distance for the entire pipeline.
 constexpr float kSunShadowMaxDistance = 64.0f;
 constexpr int kBakeSunShadowGroupSize = 16;
 constexpr int kIsoFrustumCornerCount = 8;
+
+namespace detail {
+
+struct ResolvedSun {
+    vec3 direction_ = IRRender::getSunDirection();
+    float intensity_ = IRRender::getSunIntensity();
+    float ambient_ = IRRender::getSunAmbient();
+    bool shadowsEnabled_ = IRRender::getSunShadowsEnabled();
+    bool aoEnabled_ = IRRender::getAOEnabled();
+};
+
+// A single ECS entity carrying `C_LightSource{type=DIRECTIONAL}` overrides
+// the global sun. The first three fields snapshot RenderManager state, so
+// scenes without a directional light entity still get coherent values.
+inline ResolvedSun resolveSun() {
+    ResolvedSun sun{};
+    const auto include = IREntity::getArchetype<C_LightSource>();
+    auto nodes = IREntity::queryArchetypeNodesSimple(include);
+    bool foundDirectional = false;
+    for (auto *node : nodes) {
+        auto &lights = IREntity::getComponentData<C_LightSource>(node);
+        for (int i = 0; i < node->length_; ++i) {
+            const C_LightSource &light = lights[i];
+            if (light.type_ != LightType::DIRECTIONAL)
+                continue;
+            IR_ASSERT(!foundDirectional, "Only one DIRECTIONAL light may drive the global sun");
+            foundDirectional = true;
+
+            IR_ASSERT(
+                light.direction_.z <= 0.0f,
+                "Directional light points from the world toward the sun; +Z is down, so z must be "
+                "<= 0"
+            );
+            const float length = IRMath::length(light.direction_);
+            if (length > 0.0f) {
+                sun.direction_ = light.direction_ / length;
+            }
+            sun.intensity_ = IRMath::max(0.0f, light.intensity_);
+            sun.ambient_ = IRMath::clamp(light.ambient_, 0.0f, 1.0f);
+        }
+    }
+    sun.shadowsEnabled_ = IRRender::getSunShadowsEnabled();
+    sun.aoEnabled_ = IRRender::getAOEnabled();
+    return sun;
+}
+
+} // namespace detail
 
 template <> struct System<BAKE_SUN_SHADOW_MAP> {
     struct Params {
@@ -51,9 +102,6 @@ template <> struct System<BAKE_SUN_SHADOW_MAP> {
         Buffer *sunShadowDepthMap_ = nullptr;
         Buffer *sunShadowFrameDataBuf_ = nullptr;
         Buffer *voxelFrameDataBuf_ = nullptr;
-        // Holds the BAKE-owned slice (offset useScreenSpaceShadow_ onward).
-        // Uploaded as a partial subData so the legacy prefix written by
-        // COMPUTE_SUN_SHADOW survives.
         FrameDataSun frameData_{};
     };
 
@@ -99,29 +147,10 @@ template <> struct System<BAKE_SUN_SHADOW_MAP> {
                         return;
                     }
 
-                    // BAKE-owned region: when active, upload flag + basis +
-                    // AABB; when gated off, upload just the flag (+ padding)
-                    // so COMPUTE sees the live toggle without reading stale
-                    // basis. COMPUTE overwrites the legacy prefix later this
-                    // frame — non-overlapping byte ranges.
-                    constexpr std::size_t kBakeFieldsOffset =
-                        offsetof(FrameDataSun, useScreenSpaceShadow_);
-                    constexpr std::size_t kBasisOffset =
-                        offsetof(FrameDataSun, sunBasisU_);
-                    constexpr std::size_t kFlagFieldsSize =
-                        kBasisOffset - kBakeFieldsOffset;
-                    constexpr std::size_t kBakeFieldsSize =
-                        sizeof(FrameDataSun) - kBakeFieldsOffset;
-
-                    const bool bakeActive =
-                        p->frameData_.useScreenSpaceShadow_ != 0 &&
-                        p->frameData_.shadowsEnabled_ != 0;
-                    const std::size_t uploadSize =
-                        bakeActive ? kBakeFieldsSize : kFlagFieldsSize;
-                    const auto *base = reinterpret_cast<const std::byte *>(&p->frameData_);
-                    p->sunShadowFrameDataBuf_
-                        ->subData(kBakeFieldsOffset, uploadSize, base + kBakeFieldsOffset);
-                    if (!bakeActive) {
+                    // BAKE owns the FrameDataSun UBO: upload the full struct
+                    // each frame so all downstream consumers see fresh state.
+                    p->sunShadowFrameDataBuf_->subData(0, sizeof(FrameDataSun), &p->frameData_);
+                    if (p->frameData_.shadowsEnabled_ == 0) {
                         return;
                     }
 
@@ -167,24 +196,21 @@ template <> struct System<BAKE_SUN_SHADOW_MAP> {
                         p->sunShadowFrameDataBuf_ =
                             IRRender::getNamedResource<Buffer>("ComputeSunShadowFrameData");
                     }
-                    p->frameData_.useScreenSpaceShadow_ =
-                        IRRender::getScreenSpaceShadowsEnabled() ? 1 : 0;
-                    p->frameData_.shadowsEnabled_ = IRRender::getSunShadowsEnabled() ? 1 : 0;
-                    if (p->frameData_.useScreenSpaceShadow_ == 0 ||
-                        p->frameData_.shadowsEnabled_ == 0) {
-                        return;
-                    }
 
-                    // Read sunDir directly from RenderManager rather than
-                    // resolveSun() — when an ECS directional light overrides
-                    // it, the lookup picks up the override one frame later;
-                    // baking last frame's direction is benign for static suns.
-                    vec3 sunDir = IRRender::getSunDirection();
+                    const detail::ResolvedSun sun = detail::resolveSun();
+                    vec3 sunDir = sun.direction_;
                     const float sunLen = IRMath::length(sunDir);
                     if (sunLen > 0.0f) {
                         sunDir /= sunLen;
                     }
                     p->frameData_.sunDirection_ = vec4(sunDir, 0.0f);
+                    p->frameData_.sunIntensity_ = sun.intensity_;
+                    p->frameData_.sunAmbient_ = sun.ambient_;
+                    p->frameData_.shadowsEnabled_ = sun.shadowsEnabled_ ? 1 : 0;
+                    p->frameData_.aoEnabled_ = sun.aoEnabled_ ? 1 : 0;
+                    if (p->frameData_.shadowsEnabled_ == 0) {
+                        return;
+                    }
 
                     // Pick a reference axis far from sunDir for numerical
                     // stability under near-vertical suns.
