+++ engine/render/src/ir_render.cpp
@@ -287,14 +287,6 @@ bool getSunShadowsEnabled() {
     return getRenderManager().getSunShadowsEnabled();
 }
 
-void setScreenSpaceShadowsEnabled(bool enabled) {
-    getRenderManager().setScreenSpaceShadowsEnabled(enabled);
-}
-
-bool getScreenSpaceShadowsEnabled() {
-    return getRenderManager().getScreenSpaceShadowsEnabled();
-}
-
 void setAOEnabled(bool enabled) {
     getRenderManager().setAOEnabled(enabled);
 }
